+++ front-page.php
@@ -8,11 +8,11 @@ get_header(); ?>
 	<div id="primary" class="content-area">
 		<main id="main" class="site-main" role="main">
 
-      <?php
-      the_telabotanica_module('cover-home', [
-      ]); ?>
+			<?php
+			the_telabotanica_module('cover-home');
+			?>
 
-    </main><!-- .site-main -->
-  </div><!-- .content-area -->
+		</main><!-- .site-main -->
+	</div><!-- .content-area -->
 
 <?php get_footer(); ?>