+++ modules/cover-search/cover-search.php
@@ -1,0 +1,43 @@
+<?php function telabotanica_module_cover_search($data) {
+	global $wp_query;
+	if (!isset($data->image)) $data->image = get_field('cover_image');
+	if (!isset($data->modifiers)) $data->modifiers = '';
+	?>
+	<div class="cover cover-search <?php echo $data->modifiers ?>" style="background-image: url(<?php echo $data->image['url'] ?>);">
+		<div class="layout-wrapper">
+			<div class="cover-search-content">
+			<?php
+			the_telabotanica_module('search-box', [
+				'autocomplete' => false
+			]);
+
+			if ( get_search_query() ) :
+				echo sprintf(
+					'<h1 class="cover-search-title">%s</h1>',
+					sprintf( _n(
+						'%s résultat trouvé',
+						'%s résultats trouvés',
+						$wp_query->found_posts,
+						'telabotanica'
+					), $wp_query->found_posts )
+				);
+			endif;
+			?>
+			</div>
+
+		</div>
+		<?php
+			if ( $data->image ) :
+				$credits = get_fields( $data->image['ID'] );
+				if ( $credits ) :
+					echo '<div class="cover-credits">';
+					if ($credits['link']) {
+						echo sprintf(__('%s par %s', 'telabotanica'), '<a href="' . $credits['link'] . '" target="_blank">' . $data->image['title'] . '</a>', $credits['author']);
+					} else {
+						echo sprintf(__('%s par %s', 'telabotanica'), $data->image['title'], $credits['author']);
+					}
+					echo '</div>';
+				endif;
+		endif; ?>
+	</div>
+<?php }