+++ modules/cover-search/exemples.php
@@ -1,0 +1,4 @@
+<?php
+return [
+  "Standard" => ['title' => 'Titre', 'subtitle' => 'Sous-titre'],
+];