+++ modules/search-box/search-box.php
@@ -1,29 +1,35 @@
 <?php function telabotanica_module_search_box($data) {
-	echo '<div class="search-box large">';
-		echo '<div class="search-box-wrapper">';
+	if (!isset($data->autocomplete)) $data->autocomplete = true;
+	if (!isset($data->value)) $data->value = get_search_query();
+	if (!isset($data->suggestions)) $data->suggestions = false;
+
+	echo '<div class="search-box large" data-autocomplete="' . var_export($data->autocomplete, true) . '">';
+		echo '<form role="search" method="get" action="' . esc_url( home_url( '/' ) ) . '" class="search-box-wrapper">';
 		printf(
-			'<input name="s" class="search-box-input" placeholder="%s" />',
-			__('Rechercher une plante, un projet, un mot clé...', 'telabotanica')
+			'<input name="s" class="search-box-input" placeholder="%s" value="%s" />',
+			__('Rechercher une plante, un projet, un mot clé...', 'telabotanica'),
+			$data->value
 		);
 		echo '<button type="submit" class="search-box-button">' . get_telabotanica_module('icon', ['icon' => 'search']) . '</button>';
-		echo '</div>';
+		echo '</form>';
 
-		$suggestions = ['coquelicot', 'quercus ilex', 'végétation', 'mooc'];
+		if ( $data->suggestions ) :
+			$suggestions = array_map(function($suggestion) {
+				return sprintf(
+					'<a href="%s">%s</a>',
+					'#' . $suggestion, // TODO compose URL to search results
+					$suggestion
+				);
+			}, $data->suggestions);
 
-		$suggestions = array_map(function($suggestion) {
-			return sprintf(
-				'<a href="%s">%s</a>',
-				'#' . $suggestion,
-				$suggestion
+			printf(
+				'<div class="search-box-suggestions">%s</div>',
+				sprintf(
+					__('Par exemple : %s...', 'telabotanica'),
+					implode($suggestions, ', ')
+				)
 			);
-		}, $suggestions);
+		endif;
 
-		printf(
-			'<div class="search-box-suggestions">%s</div>',
-			sprintf(
-				__('Par exemple : %s...', 'telabotanica'),
-				implode($suggestions, ', ')
-			)
-		);
 	echo '</div>';
 }