+++ CHANGELOG.md
@@ -11,7 +11,10 @@ plus de détails sur http://semver.org/lang/fr/
 
 ## x.β.α (bientôt)
 
+* ajout d'un template `search.php`
 * module `cover-home` ajouté
+* module `cover-search` ajouté
+* module `header` : support d'une version `small` pour certaines pages
 * module `search-box` ajouté
 * groupe de champs ACF "Newsletter" déplacé vers la page newsletter_compose
 * dossier `algolia` ajouté (fichiers spécifiques au plugin Algolia)